+++ CHANGELOG.md
@@ -6,6 +6,23 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+# [2.4.4] = 29 August 2024
+### Fixed
+- Prevent exception when require and/or require-dev is empty
+
+# [2.4.3] = 23 August 2024
+### Fixed
+- Fixes for dev requirements
+
+# [2.4.2] = 12 August 2024
+### Fixed
+- Changed call in Nikic parser
+
+
+# [2.4.1] = 24 June 2024
+### Fixed
+- Added JSON format to module list output
+
 # [2.4.0] = 21 June 2024
 ### Added
 - New CLI `yireo_extensionchecker:list:modules` (module name, enabled/disabled, composer version)